rework(BaseService, BaseRepo): The functionality of the basic service has been redesigned, duplication has been eliminated, and new methods for retrieving models have been added to the basic repository.

+++ app/service/base.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import Type, Generic
+from typing import Type, Generic, Optional
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -15,64 +15,50 @@ class BaseService(Generic[Repo], AService):
     async def get_all_models(
         cls,
         session: AsyncSession,
+        user_id:  Optional[int] = None,
     ) -> list[DBModel]:
         """
         Возвращает результат выполнения метода получения всех моделей пользователей из БД
         :param session: Объект сессии, полученный в качестве аргумента
         :return: Список всех моделей пользователей
         """
+        if user_id is not None:
+            return await cls.repo.get_all_by_user_id(
+                user_id=user_id,
+                session=session,
+            )
+        
         return await cls.repo.get_all(session=session)
 
     @classmethod
-    async def get_model_by_id(
+    async def get_model(
         cls,
-        model_id: int,
         session: AsyncSession,
-    ) -> DBModel:
+        user_id: int = None,
+        model_id: int = None,
+    ) -> Optional[DBModel]:
         """
-        Возвращает результат выполнения метода получения модели пользователя по ее id из БД
-        :param model_id: id модели конкретного пользователя
-        :param session: объект сессии, который получается путем выполнения зависимости (метода session_dependency объекта db_connector)
-        :return: Модель конкретного пользователя
+        Универсальный метод получения модели по user_id, model_id или обоим сразу.
+        :param session: AsyncSession
+        :param user_id: id пользователя
+        :param model_id: id модели
+        :return: Найденная модель или None
         """
-        return await cls.repo.get_by_id(
-            model_id=model_id,
-            session=session,
-        )
+        if user_id is not None and model_id is not None:
+            return await cls.repo.get_by_user_and_model_id(
+                model_id=model_id,
+                user_id=user_id,
+                session=session,
+            )
 
-    @classmethod
-    async def get_all_models_by_user_id(
-        cls,
-        user_id: int,
-        session: AsyncSession,
-    ) -> list[DBModel]:
-        """
-        Возвращает результат выполнения метода получения модели пользователя по ее id из БД
-        :param model_id: id модели конкретного пользователя
-        :param session: объект сессии, который получается путем выполнения зависимости (метода session_dependency объекта db_connector)
-        :return: Модель конкретного пользователя
-        """
-        return await cls.repo.get_all_by_user_id(
-            user_id=user_id,
-            session=session,
-        )
+        if model_id is not None:
+            return await cls.repo.get_by_id(model_id=model_id, session=session)
 
-    @classmethod
-    async def get_model_by_user_id(
-        cls,
-        user_id: int,
-        session: AsyncSession,
-    ) -> DBModel:
-        """
-        Возвращает результат выполнения метода получения модели пользователя по ее id из БД
-        :param model_id: id модели конкретного пользователя
-        :param session: объект сессии, который получается путем выполнения зависимости (метода session_dependency объекта db_connector)
-        :return: Модель конкретного пользователя
-        """
-        return await cls.repo.get_by_user_id(
-            user_id=user_id,
-            session=session,
-        )
+        if user_id is not None:
+            return await cls.repo.get_by_user_id(user_id=user_id, session=session)
+
+        return None
+
 
     @classmethod
     async def get_all_models_by_date(
@@ -96,7 +82,7 @@ class BaseService(Generic[Repo], AService):
         cls,
         scheme_in: PDScheme,
         session: AsyncSession,
-        user_id: int = None,
+        user_id: Optional[int] = None
     ) -> DBModel:
         """
         Возвращает результат выполнения метода добавления модели пользователя в БД
@@ -105,12 +91,12 @@ class BaseService(Generic[Repo], AService):
         :return: Модель пользователя, добавленную в БД
         """
         data = scheme_in.model_dump()
-        
+
         if user_id:
             data["user_id"] = user_id
 
         return await cls.repo.create(
-            scheme_in=cls.repo.model(**data),
+            model=cls.repo.model(**data),
             session=session,
         )
 
@@ -119,10 +105,10 @@ class BaseService(Generic[Repo], AService):
         cls,
         scheme_in: PDScheme,
         session: AsyncSession,
-        user_id: int = None,
-        model_id: int = None,
+        user_id: Optional[int] = None,
+        model_id: Optional[int] = None,
         partial: bool = False,
-    ) -> DBModel:
+    ) -> Optional[DBModel]:
         """
         Возвращает результат выполнения метода обновления данных модели пользователя в БД полностью или частично
         :param scheme_in: Pydantic Схема - объект, содержащий данные пользователя
@@ -131,51 +117,48 @@ class BaseService(Generic[Repo], AService):
         :param partial: Флаг, передаваем значение True или False,
         :return: Модель пользователя, обновленную в БД
         """
-        if user_id:
-            model = await cls.repo.get_by_user_id(
-                user_id=user_id,
-                session=session,
-            )
+        new_data = scheme_in.model_dump(
+            exclude_unset=partial,
+            exclude_none=True,
+        )
 
-        if model_id:
-            model = await cls.repo.get_by_id(
-                model_id=model_id,
-                session=session,
-            )
+        model = await cls.get_model(
+            session=session,
+            user_id=user_id,
+            model_id=model_id,
+        )
 
+        if not model:
+            return None
+    
         return await cls.repo.update(
-            scheme_in=scheme_in,
+            new_data=new_data,
             update_model=model,
             session=session,
-            partial=partial,
         )
 
     @classmethod
     async def delete_model(
         cls,
         session: AsyncSession,
-        user_id: int = None,
-        model_id: int = None,
-    ) -> DBModel:
+        user_id: Optional[int] = None,
+        model_id: Optional[int] = None,
+    ) -> Optional[DBModel]:
         """
         Возвращает результат выполнения метода удаления модели пользователя из БД
         :param model_id: ORM Модель - конкретный объект модели в БД, найденный по id
         :param session: Объект сессии, полученный в качестве аргумента
         :return: Модель пользователя, удаленную из БД
         """
+        model = await cls.get_model(
+            session=session,
+            user_id=user_id,
+            model_id=model_id,
+        )
 
-        if user_id:
-            model = await cls.repo.get_by_user_id(
-                user_id=user_id,
-                session=session,
-            )
-
-        if model_id:
-            model = await cls.repo.get_by_id(
-                model_id=model_id,
-                session=session,
-            )
-
+        if not model:
+            return None
+        
         return await cls.repo.delete(
             del_model=model,
             session=session,
